refactor(react,vue): improve code readability
