+++ driver/js/packages/hippy-vue/src/util/event.js
@@ -18,6 +18,41 @@
  * limitations under the License.
  */
 
+const eventHandlerType = {
+  ADD: 0,
+  REMOVE: 1,
+};
+
+const nativeEventMap = {
+  onClick: 'click',
+  onLongClick: 'longclick',
+  // onPressIn: 'touchstart', // normalization
+  // onPressOut: 'touchend', // normalization
+  onPressIn: 'pressin',
+  onPressOut: 'pressout',
+  onTouchDown: 'touchstart', // compatible with w3c standard name touchstart
+  onTouchStart: 'touchstart',
+  onTouchEnd: 'touchend',
+  onTouchMove: 'touchmove',
+  onTouchCancel: 'touchcancel',
+};
+
+const DOMEventPhase = {
+  NONE: 0,
+  CAPTURING_PHASE: 1,
+  AT_TARGET: 2,
+  BUBBLING_PHASE: 3,
+};
+
+
+function isNativeGesture(name) {
+  return !!nativeEventMap[name];
+}
+
+function translateToNativeEventName(name) {
+  return name.replace(/^(on)?/g, '').toLocaleLowerCase();
+}
+
 // event method constant
 const eventMethod = {
   ADD: 'addEventListener',
@@ -26,4 +61,9 @@ const eventMethod = {
 
 export {
   eventMethod,
+  eventHandlerType,
+  nativeEventMap,
+  DOMEventPhase,
+  isNativeGesture,
+  translateToNativeEventName,
 };
