+++ driver/js/scripts/react-configs.js
@@ -166,6 +166,12 @@ function genConfig(name) {
             declarationMap: false,
           },
           exclude: ['**/__tests__/*.test.*'],
+          include: [
+            'packages/hippy-react*/src',
+            'packages/global.d.ts',
+            'node_modules/@types/web/index.d.ts',
+            'node_modules/@types/node/index.d.ts',
+          ],
         },
       }),
     ].concat(opts.plugins || []),
