+++ package.json
@@ -16,10 +16,9 @@
     "@commitlint/config-conventional": "^16.2.1",
     "@commitlint/lint": "^16.2.1",
     "@commitlint/prompt-cli": "^16.2.1",
-    "@microsoft/api-extractor": "^7.19.4",
     "ansi-regex": "^6.0.1",
     "commitlint-plugin-function-rules": "^1.6.0",
-    "docsify-cli": "^4.4.3",
+    "docsify-cli": "^4.4.4",
     "husky": "^7.0.4",
     "is-ci": "^2.0.0",
     "markdown-it": "^12.3.2",
